Update footer and navbar links, enhance homepage layout, and integrate wallet connection functionality
